Changed most return values of the rtc_get_*() type to BYTE (since that is how most are used anyways, and gets rid of alot of casting), and fixed some rtc issues.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@24266 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/rtc/bq4830y.c
@@ -106,15 +106,15 @@ static void bq4830y_latch_write_regs(rtc_bq4830y_t *context)
     int i;
 
     context->clock_regs[BQ4830Y_REG_SECONDS & 7] &= 0x80;
-    context->clock_regs[BQ4830Y_REG_SECONDS & 7] |= (BYTE)rtc_get_second(context->latch, 1);
+    context->clock_regs[BQ4830Y_REG_SECONDS & 7] |= rtc_get_second(context->latch, 1);
     context->clock_regs[BQ4830Y_REG_MINUTES & 7] &= 0x80;
-    context->clock_regs[BQ4830Y_REG_MINUTES & 7] |= (BYTE)rtc_get_minute(context->latch, 1);
+    context->clock_regs[BQ4830Y_REG_MINUTES & 7] |= rtc_get_minute(context->latch, 1);
     context->clock_regs[BQ4830Y_REG_HOURS & 7] &= 0xc0;
-    context->clock_regs[BQ4830Y_REG_HOURS & 7] |= (BYTE)rtc_get_hour(context->latch, 1);
+    context->clock_regs[BQ4830Y_REG_HOURS & 7] |= rtc_get_hour(context->latch, 1);
     context->clock_regs[BQ4830Y_REG_DAYS_OF_WEEK & 7] &= 0xf8;
-    context->clock_regs[BQ4830Y_REG_DAYS_OF_WEEK & 7] |= (BYTE)rtc_get_weekday(context->latch) + 1;
+    context->clock_regs[BQ4830Y_REG_DAYS_OF_WEEK & 7] |= rtc_get_weekday(context->latch) + 1;
     context->clock_regs[BQ4830Y_REG_DAYS_OF_MONTH & 7] &= 0xc0;
-    context->clock_regs[BQ4830Y_REG_DAYS_OF_MONTH & 7] |= (BYTE)rtc_get_day_of_month(context->latch, 1);
+    context->clock_regs[BQ4830Y_REG_DAYS_OF_MONTH & 7] |= rtc_get_day_of_month(context->latch, 1);
     context->clock_regs[BQ4830Y_REG_MONTHS & 7] &= 0xe0;
     val = rtc_get_month(context->latch, 1);
     if (val == 7) {
@@ -123,7 +123,7 @@ static void bq4830y_latch_write_regs(rtc_bq4830y_t *context)
         val++;
     }
     context->clock_regs[BQ4830Y_REG_MONTHS & 7] |= val;
-    context->clock_regs[BQ4830Y_REG_YEARS & 7] = (BYTE)rtc_get_year(context->latch, 1);
+    context->clock_regs[BQ4830Y_REG_YEARS & 7] = rtc_get_year(context->latch, 1);
     for (i = 0; i < 8; i++) {
         context->clock_regs_changed[i] = 0;
     }
@@ -377,19 +377,19 @@ BYTE bq4830y_read(rtc_bq4830y_t *context, WORD address)
     switch (address & 0x7fff) {
         case BQ4830Y_REG_MINUTES:
             retval = context->clock_regs[address & 7] & 0x80;
-            retval |= (BYTE)rtc_get_minute(latch, 1);
+            retval |= rtc_get_minute(latch, 1);
             break;
         case BQ4830Y_REG_HOURS:
             retval = context->clock_regs[address & 7] & 0xc0;
-            retval |= (BYTE)rtc_get_hour(latch, 1);
+            retval |= rtc_get_hour(latch, 1);
             break;
         case BQ4830Y_REG_DAYS_OF_WEEK:
             retval = context->clock_regs[address & 7] & 0xf8;
-            retval |= (BYTE)rtc_get_weekday(latch) + 1;
+            retval |= rtc_get_weekday(latch) + 1;
             break;
         case BQ4830Y_REG_DAYS_OF_MONTH:
             retval = context->clock_regs[address & 7] & 0xc0;
-            retval |= (BYTE)rtc_get_day_of_month(latch, 1);
+            retval |= rtc_get_day_of_month(latch, 1);
             break;
         case BQ4830Y_REG_MONTHS:
             retval = context->clock_regs[address & 7] & 0xe0;
@@ -402,7 +402,7 @@ BYTE bq4830y_read(rtc_bq4830y_t *context, WORD address)
             retval |= (BYTE)val;
             break;
         case BQ4830Y_REG_YEARS:
-            retval = (BYTE)rtc_get_year(latch, 1);
+            retval = rtc_get_year(latch, 1);
             break;
         case BQ4830Y_REG_CONTROL:
             retval = context->clock_regs[address & 7] & 0x3f;
@@ -411,7 +411,7 @@ BYTE bq4830y_read(rtc_bq4830y_t *context, WORD address)
             break;
         case BQ4830Y_REG_SECONDS:
             retval = context->clock_halt << 7;
-            retval |= (BYTE)rtc_get_second(latch, 1);
+            retval |= rtc_get_second(latch, 1);
             break;
         default:
             retval = context->ram[address];
